Image_recognition-main/requirements.txt
absl-py==0.13.0
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
contextlib2==21.6.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.0
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-tastypie==0.14.3
djangorestframework==3.12.4
docopt==0.6.2
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gitignore==0.0.8
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
object-detection==0.0.3
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
#pypiwin32==223
pyrsistent==0.18.0
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2021.1
#pywin32==301
#pywinpty==1.1.3
#pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.3
Send2Trash==1.7.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
